Refactor AdminChronograf into UsersPage to ensure correct users via AJAX & guard
Move OrganizationsPage into /chronograf & cleanup
