+++ ui/src/admin/components/chronograf/UsersTable.js
@@ -120,10 +120,27 @@ class UsersTable extends Component {
   }
 }
 
-const {arrayOf, func, shape, string} = PropTypes
+const {arrayOf, bool, func, shape, string} = PropTypes
 
 UsersTable.propTypes = {
-  users: arrayOf(shape()),
+  users: arrayOf(
+    shape({
+      id: string,
+      links: shape({
+        self: string.isRequired,
+      }),
+      name: string.isRequired,
+      provider: string.isRequired,
+      roles: arrayOf(
+        shape({
+          name: string.isRequired,
+          organization: string.isRequired,
+        })
+      ),
+      scheme: string.isRequired,
+      superAdmin: bool,
+    })
+  ).isRequired,
   organization: shape({
     name: string.isRequired,
     id: string.isRequired,
