+++ ui/src/admin/index.js
@@ -1,5 +1,4 @@
 import AdminInfluxDBPage from './containers/AdminInfluxDBPage'
 import AdminChronografPage from './containers/AdminChronografPage'
-import OrganizationsPage from './containers/OrganizationsPage'
 
-export {AdminChronografPage, AdminInfluxDBPage, OrganizationsPage}
+export {AdminChronografPage, AdminInfluxDBPage}
